+++ packages/npm/@amazeelabs/eslint-config/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.26](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/eslint-config@1.4.25...@amazeelabs/eslint-config@1.4.26) (2022-04-21)
+
+**Note:** Version bump only for package @amazeelabs/eslint-config
+
+
+
+
+
 ## [1.4.25](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/eslint-config@1.4.24...@amazeelabs/eslint-config@1.4.25) (2022-04-14)
 
 **Note:** Version bump only for package @amazeelabs/eslint-config