+++ packages/npm/@amazeelabs/jest-preset/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.29](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/jest-preset@1.3.28...@amazeelabs/jest-preset@1.3.29) (2022-04-21)
+
+**Note:** Version bump only for package @amazeelabs/jest-preset
+
+
+
+
+
 ## [1.3.28](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/jest-preset@1.3.27...@amazeelabs/jest-preset@1.3.28) (2022-04-14)
 
 **Note:** Version bump only for package @amazeelabs/jest-preset