+++ packages/npm/@amazeelabs/recipes/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.14.1](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/recipes@1.14.0...@amazeelabs/recipes@1.14.1) (2022-04-21)
+
+**Note:** Version bump only for package @amazeelabs/recipes
+
+
+
+
+
 # [1.14.0](https://github.com/AmazeeLabs/silverback-mono/compare/@amazeelabs/recipes@1.13.4...@amazeelabs/recipes@1.14.0) (2022-04-19)