+++ packages/npm/verdaccio-git/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.58](https://github.com/AmazeeLabs/storage-git/compare/verdaccio-git@0.1.57...verdaccio-git@0.1.58) (2022-04-21)
+
+**Note:** Version bump only for package verdaccio-git
+
+
+
+
+
 ## [0.1.57](https://github.com/AmazeeLabs/storage-git/compare/verdaccio-git@0.1.56...verdaccio-git@0.1.57) (2022-04-14)
 
 **Note:** Version bump only for package verdaccio-git